Connect to Outlook Business 365

These functions create a connection to Outlook Business 365 and saves the connection to the certemail package environment. The get_*() functions retrieve properties from this connection.

connect_outlook365(tenant = read_secret("mail.tenant"), error_on_fail = FALSE)

get_name(account = connect_outlook365())

get_job_title(account = connect_outlook365())

get_department(account = connect_outlook365())

get_address(account = connect_outlook365())

get_phone_numbers(account = connect_outlook365())

get_mail_address(account = connect_outlook365())

get_inbox_name(account = connect_outlook365())

get_drafts_name(account = connect_outlook365())

get_certe_name(user = Sys.info()["user"], account = connect_outlook365())

get_certe_name_and_job_title(
  user = Sys.info()["user"],
  account = connect_outlook365()
)

get_certe_location(account = connect_outlook365())

get_certe_signature(account = connect_outlook365(), plain = FALSE)

Arguments

tenant

the tenant to use for Microsoft365R::get_business_outlook()

error_on_fail

a logical to indicate whether an error must be thrown if no connection can be made

account

a Microsoft 365 account to use for looking up properties. This has to be an object as returned by connect_outlook365() or Microsoft365R::get_business_outlook().

user

Certe user ID, to look up the secret user.{user}.fullname

plain

a logical to indicate whether textual formatting should not be applied

Details

The get_certe_name() and get_certe_name_and_job_title() functions first look for the secret user.{user}.fullname and falls back to get_name() if it is not available.