Connect to Calendar in Outlook 365

After connecting using connect_outlook(), the get_*() functions retrieve properties from this connection.

calendar_get_events(
  start_date = Sys.Date(),
  end_date = start_date,
  account = connect_outlook()
)

calendar_unavailable_users(
  start_date = Sys.Date(),
  end_date = start_date,
  account = connect_outlook()
)

Arguments

start_date, end_date

date range

account

the Microsoft 365 Outlook account object, e.g. as returned by connect_outlook()

Details

calendar_get_events() retrieves a data.frame with all appointments from the shared calendar set in account.

calendar_unavailable_users() returns a vector of appointment subjects that are shown as "free" within start_date and end_date.