Download Email Attachments Using Microsoft 365

This uses the Microsoft365R package to download email attachments via Microsoft 365. Connection will be made internally using get_business_outlook() using the Certe tenant ID.

download_mail_attachment(
  path = getwd(),
  search = "hasattachment:yes",
  search_subject = NULL,
  search_from = NULL,
  search_when = NULL,
  folder = get_inbox_name(),
  n = 50,
  sort = "received desc",
  overwrite = TRUE
)

Arguments

path
location to save attachment(s) in

search
an ODATA filter, ignores sort and defaults to search only mails with attachments

search_subject
a character, equal to search = "subject:(search_subject)", case-insensitive

search_from
a character, equal to search = "from:(search_from)", case-insensitive

search_when
a Date vector of size 1 or 2, equal to search = "received:date1..date2", see Examples

folder
email folder to search in, defaults to Inbox name of the current user by calling get_inbox_name()

n
maximum number of emails to list

sort
initial sorting

overwrite
logical to indicate whether existing local files should be overwritten

Details

If both search_subject and search_from are provided, they will be matched as 'AND'. search_from can contain any sender name or email address. If search_when has a length over 2, the first and last value will be taken.

See this page for all filtering options using the search argument: https://support.microsoft.com/en-gb/office/how-to-search-in-outlook-d824d1e9-a255-4c8a-8553-276fb895a8da.

See also

Examples

if (FALSE) {
download_mail_attachment(search_when = "2021-10-28")

if (require("certetoolbox")) {
  download_mail_attachment(search_from = "glims", search_when = last_week())

  download_mail_attachment(search_from = "drenthe", search_when = yesterday(), n = 1)
}
}