Send emails using Microsoft 365

Send emails using Microsoft 365

mail(
  body,
  subject,
  to,
  cc = NULL,
  bcc = read_secret("mail.auto_bcc"),
  reply_to = NULL,
  attachment = NULL,
  header = FALSE,
  footer = FALSE,
  background = colourpicker("certeblauw3"),
  send = TRUE,
  md_body = TRUE,
  signature = TRUE,
  signature_name = get_name_and_job_title(),
  signature_address = get_mail_address(),
  automated_notice = !interactive(),
  outlook = FALSE,
  save_location = read_secret("mail.export_path"),
  expect = NULL,
  ...
)

mail_plain(
  body,
  subject,
  to,
  cc = NULL,
  bcc = read_secret("mail.auto_bcc"),
  reply_to = read_secret("mail.auto_reply_to"),
  attachment = NULL,
  send = TRUE,
  ...
)

mail_dataset(df, font.size = 11, ..., plain = FALSE)

mail_image(image_path, width = NULL, ...)

mail_on_error(expr, to = read_secret("mail.error_to"), ...)

download_mail_attachment(
  path = getwd(),
  search = NULL,
  folder = "Postvak IN",
  n = 50,
  sort = "received desc",
  overwrite = TRUE
)

Arguments

body
Tekst van de e-mail.

subject
Onderwerp van de e-mail. Spaties aan begin en einde worden weggesneden.

to
Geaddresseerde (veld 'Aan') van de e-mail. Kan ook vector zijn. Standaard het e-mailadres zoals gedefinieerd in de omgevingsvariabele van de ingelogde gebruiker.

cc
Geaddresseerde (veld 'CC') van de e-mail. Kan ook vector zijn.

bcc
Geaddresseerde (veld 'BCC') van de e-mail. Kan ook vector zijn. Adressen die voorkomen in to of cc worden eruit gehaald. Als in to of cc minimaal 1 data-onderzoeker staat, wordt dit standaard niet naar de data-onderzoekers gestuurd als BCC.

reply_to
Geaddresseerde (veld 'Van') van de e-mail. In Outlook moeten de rechten vastgelegd zijn om vanaf dit account te versturen.

attachment
Bijlagen: vector van locaties. Als de locatie niet bestaat, wordt er een fout gegeven en wordt de mail niet verstuurd.

header, footer
Extra info voor boven en onder de body. Zie blocks voor het maken van blokken die hier geplaatst kunnen worden. Bij FALSE worden deze blokken verborgen.

background
Achtergrond van de mail (standaard: colourpicker("certeblauw3")). Gebruik "", NULL of FALSE om geen achtergrond te gebruiken.

send
Direct versturen van de e-mail. Met send = FALSE wordt de e-mail op het scherm weergegeven. Standaard TRUE.

md_body
Tekst van de e-mail als Markdown opmaken (standaard: TRUE).

signature
Handtekening aan e-mail toevoegen (standaard: TRUE).

signature_name
Naam die onder 'Met vriendelijke groet' komt te staan.

signature_address
E-mailaderes die onder 'Met vriendelijke groet' komt te staan.

automated_notice
Onderaan de mail deze tekst weergeven: "Deze mail is geautomatiseerd verstuurd." (standaard: TRUE).

outlook
Mail opstellen met Outlook (standaard: FALSE). Anders wordt de mail rechtstreeks via de SMTP-server verzonden.

save_location
De maplocatie waar de mail opgeslagen wordt nadat deze verzonden is. Deze wordt opgeslagen als .rds-bestand (met evt. bijlagen), die in RStudio geopend kan worden en geprint kan worden in de Console voor alle maildetails. De mail kan dan zelfs opnieuw verstuurd worden. Gebruik "", NULL of FALSE om de mail niet op te slaan.

expect
Standaard is leeg. Een expressie die TRUE moet zijn alvorens te mailen. Als dit FALSE is om een fout oplevert, wordt mail_on_error() gedraaid met details over de fout.

...
Voor mail_on_error: parameters die doorgegeven worden aan mail().
Voor mail_dataset(): parameters die doorgegeven worden aan tbl_flextable().

df
Een dataset

font.size
Standaard is 11. Lettergrootte die doorgegeven wordt aan tbl_flextable().

image_path
path of image

width
required width of image, must be in CSS style such as "200px"

expr
Expressie die gedraaid moet worden

path
save location for attachment(s)

search
an ODATA filter, ignores sort

folder
email folder to search in

n
maximum number of email to list

sort
initial sorting

overwrite
logical to indicate whether existing local files should be overwritten

Details

De functie mail_on_error() is speciaal voor gebruik in geautomatiseerde scripts. Het evalueert de input en verstuurt een mail met de expressie wanneer deze een fout retourneert.

De functies mail_dataset() en mail_image() zijn om resp. een tabel met tbl_flextable() en een afbeelding in de mail te plaatsen.