Send emails using Microsoft 365

This uses the Microsoft365R package to send email via Microsoft 365. Connection will be made internally using Microsoft365R::get_business_outlook() using the Certe tenant ID.

mail(
  body,
  subject,
  to,
  cc = NULL,
  bcc = read_secret("mail.auto_bcc"),
  reply_to = NULL,
  attachment = NULL,
  header = FALSE,
  footer = FALSE,
  background = colourpicker("certeblauw6"),
  send = TRUE,
  markdown = TRUE,
  signature = TRUE,
  signature_name = get_name_and_job_title(),
  signature_address = get_mail_address(),
  automated_notice = !interactive(),
  save_location = read_secret("mail.export_path"),
  expect = NULL,
  ...
)

mail_plain(
  body,
  subject,
  to,
  cc = NULL,
  bcc = read_secret("mail.auto_bcc"),
  reply_to = read_secret("mail.auto_reply_to"),
  attachment = NULL,
  send = TRUE,
  ...
)

mail_image(image_path, width = NULL, ...)

mail_on_error(expr, to = read_secret("mail.error_to"), ...)

download_mail_attachment(
  path = getwd(),
  search = NULL,
  search_subject = NULL,
  search_from = NULL,
  search_when = NULL,
  folder = get_inbox_name(),
  n = 50,
  sort = "received desc",
  overwrite = TRUE
)

Arguments

body
body of email, allows markdown if markdown = TRUE

subject
subject of email

to
field 'to', can be character vector

cc
field 'CC', can be character vector

bcc
field 'BCC', can be character vector

reply_to
field 'reply-to'

attachment
character (vector) of file location(s)

header, footer
extra text for header or footer, allows markdown if markdown = TRUE. See blastula::blocks() to build blocks for these sections.

background
background of the surrounding area in the email. Use "", NULL or FALSE to remove background.

send
directly send email, FALSE will show the email in the Viewer pane and will ask whether the email should be saved to the Drafts folder of the current Microsoft 365 user.

markdown
treat body, header and footer as markdown

signature
add signature to email

signature_name
signature name

signature_address
signature email address, placed directly below signature_name

automated_notice
print a notice that the mail was send automatically (default is TRUE is not in interactive() mode)

save_location
location to save email object to, which consists of all email details and can be printed in the R console

expect
expression which should return TRUE prior to sending the email

...
arguments for mail()

image_path
path of image

width
required width of image, must be in CSS style such as "200px" or "100%"

expr
expression to test, an email will be sent if this expression returns an error

path
location to save attachment(s) in

search
an ODATA filter, ignores sort

search_subject
a character, equal to search = "subject:(search_subject)", case-insensitive

search_from
a character, equal to search = "from:(search_from)", case-insensitive

search_when
a Date vector of size 1 or 2, equal to search = "received:date1..date2", see Examples

folder
email folder to search in, defaults to Inbox name of the current user by calling get_inbox_name()

n
maximum number of emails to list

sort
initial sorting

overwrite
logical to indicate whether existing local files should be overwritten

Details

mail_on_error() can be used for automated scripts.

mail_plain() send a plain email, without markdown support and with no signature.

If both search_subject and search_from are provided, they will be matched as 'AND'. search_from can contain any sender name or email address. If search_when has a length over 2, the first and last value will be taken.

Examples

if (FALSE) {
download_mail_attachment(search_when = "2021-10-28")

library(certetoolbox)
download_mail_attachment(search_from = "glims", search_when = last_week())

download_mail_attachment(search_from = "drenthe", search_when = yesterday(), n = 1)
}