Send Emails Using Microsoft 365

These funnctions use the Microsoft365R R package to send emails via Microsoft 365. They require an Outlook Business account.

mail(
  body,
  subject = "",
  to = NULL,
  cc = read_secret("mail.auto_cc"),
  bcc = read_secret("mail.auto_bcc"),
  reply_to = read_secret("mail.auto_reply_to"),
  attachment = NULL,
  header = FALSE,
  footer = FALSE,
  background = certestyle::colourpicker("certeblauw6"),
  send = TRUE,
  markdown = TRUE,
  signature = get_certe_signature(account = account),
  automated_notice = !interactive(),
  save_location = read_secret("mail.export_path"),
  expect = NULL,
  account = connect_outlook365(),
  ...
)

mail_plain(
  body,
  subject = "",
  to = NULL,
  cc = read_secret("mail.auto_cc"),
  bcc = read_secret("mail.auto_bcc"),
  reply_to = read_secret("mail.auto_reply_to"),
  attachment = NULL,
  send = TRUE,
  ...
)

mail_image(image_path, width = NULL, ...)

mail_on_error(expr, to = read_secret("mail.error_to"), ...)

Arguments

body

body of email, allows markdown if markdown = TRUE

subject

subject of email

to

field 'to', can be character vector

cc

field 'CC', can be character vector

bcc

field 'BCC', can be character vector

reply_to

field 'reply-to'

attachment

character (vector) of file location(s)

header, footer

extra text for header or footer, allows markdown if markdown = TRUE. See blastula::blocks() to build blocks for these sections.

background

background of the surrounding area in the email. Use "", NULL or FALSE to remove background.

send

directly send email, FALSE will show the email in the Viewer pane and will ask whether the email should be saved to the Drafts folder of the current Microsoft 365 user.

markdown

treat body, header and footer as markdown

signature

text to print as email signature, or NULL to omit it, defaults to get_certe_signature()

automated_notice

print a notice that the mail was send automatically (default is TRUE is not in interactive() mode)

save_location

location to save email object to, which consists of all email details and can be printed in the R console

expect

expression which should return TRUE prior to sending the email

account

a Microsoft 365 account to use for sending the mail. This has to be an object as returned by connect_outlook365() or Microsoft365R::get_business_outlook(). Using account = FALSE is equal to setting send = FALSE.

...

arguments for mail()

image_path

path of image

width

required width of image, must be in CSS style such as "200px" or "100%"

expr

expression to test, an email will be sent if this expression returns an error

Details

mail_on_error() can be used for automated scripts.

mail_plain() sends a plain email, without markdown support and with no signature.

See also

Examples

mail("test123", "test456", to = "mail@domain.com", account = NULL)
#> No valid Microsoft 365 account set with argument `account`, forcing `send = FALSE`
#> Warning: In read_secret(): environmental variable 'secrets_file' not set
#> Warning: In read_secret(): environmental variable 'secrets_file' not set
#> Warning: In read_secret(): environmental variable 'secrets_file' not set
#> Warning: In read_secret(): environmental variable 'secrets_file' not set
#> Details of email creation:
#> Date/time: 3 mei 2022 12:01:00
#> Subject:   test456
#> Reply to:  
#> To:        mail@domain.com
#> CC:        
#> BCC:       

mail_plain("test123", "test456", to = "mail@domain.com", account = NULL)
#> No valid Microsoft 365 account set with argument `account`, forcing `send = FALSE`
#> Warning: In read_secret(): environmental variable 'secrets_file' not set
#> Warning: In read_secret(): environmental variable 'secrets_file' not set
#> Warning: In read_secret(): environmental variable 'secrets_file' not set
#> Warning: In read_secret(): environmental variable 'secrets_file' not set
#> Details of email creation:
#> Date/time: 3 mei 2022 12:01:00
#> Subject:   test456
#> Reply to:  
#> To:        mail@domain.com
#> CC:        
#> BCC:       

mail(mail_image(image_path = system.file("test.jpg", package = "certemail")),
    "test456", to = "mail@domain.com", account = NULL)
#> No valid Microsoft 365 account set with argument `account`, forcing `send = FALSE`
#> Warning: In read_secret(): environmental variable 'secrets_file' not set
#> Warning: In read_secret(): environmental variable 'secrets_file' not set
#> Warning: In read_secret(): environmental variable 'secrets_file' not set
#> Warning: In read_secret(): environmental variable 'secrets_file' not set
#> Details of email creation:
#> Date/time: 3 mei 2022 12:01:00
#> Subject:   test456
#> Reply to:  
#> To:        mail@domain.com
#> CC:        
#> BCC:       

# data.frames will be transformed with certestyle::plain_html_table()
mail(mtcars[1:5, ],
     subject = "Check these cars!",
     to = "somebody@domain.org",
     account = FALSE)
#> Warning: In read_secret(): environmental variable 'secrets_file' not set
#> Warning: In read_secret(): environmental variable 'secrets_file' not set
#> Warning: In read_secret(): environmental variable 'secrets_file' not set
#> Warning: In read_secret(): environmental variable 'secrets_file' not set
#> Details of email creation:
#> Date/time: 3 mei 2022 12:01:00
#> Subject:   Check these cars!
#> Reply to:  
#> To:        somebody@domain.org
#> CC:        
#> BCC: